pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "leilao-inteligente"
version = "0.1.0"
description = "Sistema de IA para monitoramento e analise de leiloes de gado"
readme = "README.md"
requires-python = ">=3.12"
license = "MIT"
authors = [
    { name = "Mority" },
]
dependencies = [
    "yt-dlp>=2024.0.0",
    "opencv-python-headless>=4.9.0",
    "google-generativeai>=0.8.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "sqlalchemy>=2.0.30",
    "alembic>=1.13.0",
    "typer[all]>=0.12.0",
    "rich>=13.7.0",
    "python-dotenv>=1.0.0",
    "numpy>=1.26.0",
    "Pillow>=10.3.0",
]

[project.optional-dependencies]
dashboard = [
    "streamlit>=1.35.0",
    "plotly>=5.22.0",
    "pandas>=2.2.0",
]
dev = [
    "pytest>=8.2.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

[project.scripts]
leilao = "leilao_inteligente.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/leilao_inteligente"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "SIM", "RUF"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
